Adicionado metodo cliente por id

+++ backend/src/main/java/com/ricardo/backend/repositorios/ClienteRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ricardo.backend.entidades.Cliente;
 
 // Camada de acesso ao banco de dados
-public interface ClienteRepositorio extends JpaRepository<Cliente, Long>{
+public interface ClienteRepositorio extends JpaRepository<Cliente, Integer>{
 
 }
